+++ doc/changelog.md
@@ -3,6 +3,7 @@
 - (fix) support for HTTP/2 (thanks to @Dean-Corso for the report)
 - (fix) displaying correct title for custom Youtube feeds
 - (etc) do not auto-refresh feeds on startup
+- (etc) show current feed/folder in search placeholder
 
 # v2.7 (2026-06-24)